A little refactor in public types; fix tests
